extracted api responses to a different file
